web/requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.6.0
billiard==3.6.4.0
branca==0.4.2
celery==5.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-colorfield==0.7.2
django-environ==0.8.1
django-filter==21.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
flake8==4.0.1
folium==0.12.1
idna==3.3
Jinja2==3.0.2
kombu==5.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.4
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2==2.8.6
pycodestyle==2.8.0
pyflakes==2.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.26.0
rope==0.21.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
